revert generated code to default in order to create patch

+++ lib/Cpp-rest-openapi-client/include/CppRestOpenAPIClient/api/DeviceApi.h
@@ -58,7 +58,7 @@ public:
     pplx::task<std::vector<std::shared_ptr<Message>>> listCommands(
         utility::string_t companyName,
         utility::string_t carName,
-        boost::optional<int64_t> since,
+        boost::optional<int32_t> since,
         boost::optional<bool> wait
     ) const;
     /// <summary>
@@ -74,7 +74,7 @@ public:
     pplx::task<std::vector<std::shared_ptr<Message>>> listStatuses(
         utility::string_t companyName,
         utility::string_t carName,
-        boost::optional<int64_t> since,
+        boost::optional<int32_t> since,
         boost::optional<bool> wait
     ) const;
     /// <summary>
